requirements.txt
discord.py
python-dotenv
google-genai
aiohttp